// frontend/src/api/settings.ts
export const settingsKeys = { all: ['settings'] as const }
// ...
export interface Signatory {
  role: string
  name: string
  title: string
  image: string
  has_image?: boolean
}